Small README improvements and typo fixes
**Problems**

This PR seeks to improve the README phrasing and fixes some small phrasing typos. I hope this is helpful and not too nitpicky :)

I removed some uses of the word "simple," as it doesn't provide any value and "simple" is always relative; using this framework would not be simple to someone unfamiliar with the delegate pattern for example.

**Solution**

I've updated the README with minor improvements.

**Testing**

No testing is necessary beyond a PR review.

GitHub Pull Request: #2
https://github.com/twitter/TwitterTextEditor/pull/2

JIRA Issues: IOS-94620

Differential Revision: https://phabricator.twitter.biz/D610561