Change Root.plist in Settings.bundle to have default value NO
**Problems**

In the current implementation, the value retrieved from UserDefaults is not True even if it appears True in Settings.app.

**Solution**

I suggested this #3 . We discussed and decided that it would be simpler to set the default value to `NO`, so I changed the default value in Settings.bundle to `NO`.

**Testing**

I tested this on my actual device iPad Pro.

GitHub Pull Request: #11
https://github.com/twitter/TwitterTextEditor/pull/11

JIRA Issues: IOS-94620

Differential Revision: https://phabricator.twitter.biz/D610563