[TwitterTextEditor] FIX: Missing force unwrap on `textContentType`
**Problems**

`textContentType` is declared as `null_unspecified` which is basically
nullable. In Swift, such property is declared as `Type!` and forcibly
unwrapped when it is read as Swift type and crash if it is `null`.
(So, this is very questionable Objective-C to Swift conversion, because
this property's default value is actually, `nil`.)

Since `TextEditorView`'s `textContentType` is simply proxy to
underlying `UITextView`, it should be also `UITextContentType!`,
instead of `UITextContentType`, or forcibly unwrapping is happening
and crash the app.

**Solution**

Change `UITextContentType` to `UITextContentType!`.

JIRA Issues: IOS-98409

Differential Revision: https://phabricator.twitter.biz/D651012